feat: extract media directive handlers
